Reorganised cli/main.py
